Moved all attributes from the testdata assembly to the Hook tests. Splitted them to one attribute per file in order to have a better overview which attributes already exist to prevent code duplication.
