+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateAfterTestHookAttribute.cs
@@ -1,0 +1,14 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
+    public sealed class ActivateAfterTestHookAttribute : ExecutionHookAttribute
+    {
+        public override void AfterTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateAfterTestHookAttribute));
+        }
+    }
+}
