+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateAfterTestHookThrowingExceptionAttribute.cs
@@ -1,0 +1,15 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public sealed class ActivateAfterTestHookThrowingExceptionAttribute : ExecutionHookAttribute
+    {
+        public override void AfterTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateAfterTestHookThrowingExceptionAttribute));
+            throw new Exception("After test hook crashed!!");
+        }
+    }
+}
