+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateBeforeTestHookAttribute.cs
@@ -1,0 +1,14 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
+    public sealed class ActivateBeforeTestHookAttribute : ExecutionHookAttribute
+    {
+        public override void BeforeTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateBeforeTestHookAttribute));
+        }
+    }
+}
