+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateBeforeTestHookThrowingExceptionAttribute.cs
@@ -1,0 +1,15 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public sealed class ActivateBeforeTestHookThrowingExceptionAttribute : ExecutionHookAttribute
+    {
+        public override void BeforeTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateBeforeTestHookThrowingExceptionAttribute));
+            throw new Exception("Before test hook crashed!!");
+        }
+    }
+}
