+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateClassLevelAfterTestHooksAttribute.cs
@@ -1,0 +1,14 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public sealed class ActivateClassLevelAfterTestHooksAttribute : ExecutionHookAttribute
+    {
+        public override void AfterTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateClassLevelAfterTestHooksAttribute));
+        }
+    }
+}
