+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateClassLevelBeforeTestHooksAttribute.cs
@@ -1,0 +1,14 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public sealed class ActivateClassLevelBeforeTestHooksAttribute : ExecutionHookAttribute
+    {
+        public override void BeforeTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateClassLevelBeforeTestHooksAttribute));
+        }
+    }
+}
