+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateLongRunningBeforeTestHookAttribute.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Threading;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public sealed class ActivateLongRunningBeforeTestHookAttribute : ExecutionHookAttribute
+    {
+        public override void BeforeTestHook(TestExecutionContext context)
+        {
+            Thread.Sleep(500);
+            TestLog.LogMessage(nameof(ActivateLongRunningBeforeTestHookAttribute));
+        }
+    }
+}
