+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateMethodLevelBeforeTestHooksAttribute.cs
@@ -1,0 +1,14 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public sealed class ActivateMethodLevelBeforeTestHooksAttribute : ExecutionHookAttribute
+    {
+        public override void BeforeTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(nameof(ActivateMethodLevelBeforeTestHooksAttribute));
+        }
+    }
+}
