+++ src/NUnitFramework/tests/ExecutionHooks/Common/ActivateTestHookAttribute.cs
@@ -1,0 +1,19 @@
+using System;
+using NUnit.Framework.Internal;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method)]
+    public sealed class ActivateTestHookAttribute : ExecutionHookAttribute
+    {
+        public override void BeforeTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(HookIdentifiers.BeforeTestHook + $"({context.CurrentTest.MethodName})");
+        }
+
+        public override void AfterTestHook(TestExecutionContext context)
+        {
+            TestLog.LogMessage(HookIdentifiers.AfterTestHook + $"({context.CurrentTest.MethodName})");
+        }
+    }
+}
