+++ src/NUnitFramework/tests/ExecutionHooks/Common/LogTestActionAttribute.cs
@@ -1,0 +1,21 @@
+using System;
+using NUnit.Framework.Interfaces;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
+    public class LogTestActionAttribute : Attribute, ITestAction
+    {
+        public void BeforeTest(ITest test)
+        {
+            TestLog.LogCurrentMethodWithContextInfo(test.IsSuite ? "Suite" : "Test");
+        }
+
+        public void AfterTest(ITest test)
+        {
+            TestLog.LogCurrentMethodWithContextInfo(test.IsSuite ? "Suite" : "Test");
+        }
+
+        public ActionTargets Targets => ActionTargets.Test | ActionTargets.Suite;
+    }
+}
