+++ src/NUnitFramework/tests/ExecutionHooks/Common/SimpleTestActionAttribute.cs
@@ -1,0 +1,24 @@
+using System;
+using NUnit.Framework.Interfaces;
+
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public sealed class SimpleTestActionAttribute : Attribute, ITestAction
+    {
+        public static readonly string LogStringForBeforeTest = $"{nameof(SimpleTestActionAttribute)}.{nameof(BeforeTest)}";
+        public static readonly string LogStringForAfterTest = $"{nameof(SimpleTestActionAttribute)}.{nameof(AfterTest)}";
+
+        public void BeforeTest(ITest test)
+        {
+            TestLog.LogMessage(LogStringForBeforeTest);
+        }
+
+        public void AfterTest(ITest test)
+        {
+            TestLog.LogMessage(LogStringForAfterTest);
+        }
+
+        public ActionTargets Targets => ActionTargets.Suite;
+    }
+}
