+++ src/NUnitFramework/tests/ExecutionHooks/Common/TestLog.cs
@@ -4,9 +4,9 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Threading;
 
-namespace NUnit.TestData.ExecutionHookTests
+namespace NUnit.Framework.Tests.ExecutionHooks.Common
 {
-    public static class TestLog
+    internal static class TestLog
     {
         private static readonly AsyncLocal<List<string>> LocalLogs = new();
         private static readonly object LogLock = new();
