+++ src/NUnitFramework/tests/ExecutionHooks/ExceptionHandling/BeforeTestHookThrowsExceptionTestStopsAfterTestHookExecutesTests.cs
@@ -1,8 +1,8 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
 using NUnit.Framework.Internal;
+using NUnit.Framework.Tests.ExecutionHooks.Common;
 using NUnit.Framework.Tests.TestUtilities;
-using NUnit.TestData.ExecutionHookTests;
 
 namespace NUnit.Framework.Tests.ExecutionHooks.ExceptionHandling
 {
