+++ src/NUnitFramework/tests/ExecutionHooks/Threading/SynchronousHookInvocationTests.cs
@@ -3,8 +3,8 @@
 using System;
 using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
+using NUnit.Framework.Tests.ExecutionHooks.Common;
 using NUnit.Framework.Tests.TestUtilities;
-using NUnit.TestData.ExecutionHookTests;
 
 namespace NUnit.Framework.Tests.ExecutionHooks.ThreadingTests
 {
